+++ lib/scanner/plugins/ssl/sweet32.rb
@@ -6,6 +6,12 @@ module Yawast
           def self.get_tdes_session_msg_count(uri)
             # this method will send a number of HEAD requests to see
             #  if the connection is eventually killed.
+            unless check_tdes(uri)
+              #if the OpenSSL install doesn't support 3DES, bailout
+              Yawast::Utilities.puts_error "Your copy of OpenSSL doesn't support 3DES cipher suites - SWEET32 test aborted."
+              return
+            end
+
             puts 'TLS Session Request Limit: Checking number of requests accepted using 3DES suites...'
 
             count = 0
@@ -79,6 +85,50 @@ module Yawast
 
             puts
             Yawast::Utilities.puts_vuln 'TLS Session Request Limit: Connection not terminated after 10,000 requests; possibly vulnerable to SWEET32'
+          end
+
+          def self.check_tdes(uri)
+            puts 'Confirming your OpenSSL supports 3DES cipher suites...'
+
+            dns = Resolv::DNS.new
+
+            if IPAddress.valid? uri.host
+              ip = IPAddress.parse uri.host
+            else
+              ip = dns.getaddresses(uri.host)[0]
+            end
+
+            #find all versions that don't include '_server' or '_client'
+            versions = OpenSSL::SSL::SSLContext::METHODS.find_all { |v| !v.to_s.include?('_client') && !v.to_s.include?('_server')}
+
+            versions.each do |version|
+              #ignore SSLv23, as it's an auto-negotiate, which just adds noise
+              if version.to_s != 'SSLv23' && version.to_s != 'SSLv2'
+                #try to get the list of ciphers supported for each version
+                ciphers = nil
+
+                get_ciphers_failed = false
+                begin
+                  ciphers = OpenSSL::SSL::SSLContext.new(version).ciphers
+                rescue => e
+                  Yawast::Utilities.puts_error "\tError getting cipher suites for #{version}, skipping. (#{e.message})"
+                  get_ciphers_failed = true
+                end
+
+                if ciphers != nil
+                  ciphers.each do |cipher|
+                    if cipher[0].include?('3DES') || cipher[0].include?('CBC3')
+                      return true
+                    end
+                  end
+                elsif !get_ciphers_failed
+                  Yawast::Utilities.puts_info "\t#{version}: No cipher suites available."
+                end
+              end
+            end
+
+            puts ''
+            return false
           end
         end
       end