+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## 0.5.0 - In Development
 
 * [#75](https://github.com/adamcaudill/yawast/issues/75) - Use internal SSL scanner for non-standard ports
+* [#84](https://github.com/adamcaudill/yawast/issues/84) - Improve the display of ct_precert_scts
 * [#76](https://github.com/adamcaudill/yawast/issues/76) - Bug: Handle error for OpenSSL version support error
+* Various code and other improvements.
 
 ## 0.4.0 - 2016-11-03