+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.5.0 - In Development
 
+* [#53](https://github.com/adamcaudill/yawast/issues/53) - Added checks for .well-known URLs
 * [#75](https://github.com/adamcaudill/yawast/issues/75) - Use internal SSL scanner for non-standard ports
 * [#84](https://github.com/adamcaudill/yawast/issues/84) - Improve the display of ct_precert_scts
 * [#86](https://github.com/adamcaudill/yawast/issues/86) - Add check for Tomcat Manager & common passwords