+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * [#76](https://github.com/adamcaudill/yawast/issues/76) - Bug: Handle error for OpenSSL version support error
 * [#98](https://github.com/adamcaudill/yawast/issues/98) - Bug: SWEET32 Test Fails if 3DES Not Support By Latest Server Supported TLS Version
 * [#99](https://github.com/adamcaudill/yawast/issues/99) - Bug: Cloudflare SWEET32 False Positive
+* [#101](https://github.com/adamcaudill/yawast/issues/101) - Bug: SWEET32 False Negative
 * Various code and other improvements.
 
 ## 0.4.0 - 2016-11-03